+++ src/tyr/route_serializer.cc
@@ -11,6 +11,8 @@
 #include "tyr/serializers.h"
 
 #include <valhalla/proto/directions_options.pb.h>
+#include <valhalla/proto/tripdirections.pb.h>
+#include <valhalla/proto/trippath.pb.h>
 
 using namespace valhalla;
 using namespace valhalla::tyr;
@@ -75,7 +77,7 @@ namespace valhalla {
 namespace tyr {
 
 std::string serializeDirections(const valhalla_request_t& request,
-                                const std::list<TripPath>& path_legs,
+                                std::list<TripPath>& path_legs,
                                 const std::list<TripDirections>& directions_legs) {
   // serialize them
   switch (request.options.format()) {